adjusting conditions on when to use counts only consumer
