+++ server/src/main/java/org/elasticsearch/action/search/SearchPhaseController.java
@@ -793,15 +793,17 @@ public final class SearchPhaseController {
         Consumer<Exception> onPartialMergeFailure
     ) {
         final int size = request.source() == null || request.source().size() == -1 ? SearchService.DEFAULT_SIZE : request.source().size();
-        // Use the new CountOnlyQueryPhaseResultConsumer for requests without aggs, suggest, etc. things only wanting a total count and
+        // Use CountOnlyQueryPhaseResultConsumer for requests without aggs, suggest, etc. things only wanting a total count and
         // returning no hits
         if (size == 0
-            && request.source().aggregations() == null
-            && request.source().suggest() == null
-            && request.source().knnSearch().isEmpty()
-            && request.source().profile() == false
+            && (request.source() == null
+                || (request.source().aggregations() == null
+                    && request.source().suggest() == null
+                    && request.source().rankBuilder() == null
+                    && request.source().knnSearch().isEmpty()
+                    && request.source().profile() == false))
             && request.resolveTrackTotalHitsUpTo() == SearchContext.TRACK_TOTAL_HITS_ACCURATE) {
-            return new QueryPhaseResultConsumer.CountOnlyQueryPhaseResultConsumer(listener, numShards);
+            return new CountOnlyQueryPhaseResultConsumer(listener, numShards);
         }
         return new QueryPhaseResultConsumer(
             request,
